misc tensorflow test fixes for newer versions of numpy + matplotlib

+++ src/probflow/models/discrete_model.py
@@ -105,7 +105,7 @@ class DiscreteModel(ContinuousModel):
             samples = samples.reshape([Ns, N])
 
         # Plot the predictive distributions
-        rows = np.ceil(N / cols)
+        rows = int(np.ceil(N / cols))
         for i in range(N):
             plt.subplot(rows, cols, i + 1)
             plot_discrete_dist(samples[:, i])
